* r502 release sp build-18890

+++ client/UserConnection.cpp
@@ -433,7 +433,7 @@ void UserConnection::updateChunkSize(int64_t leafSize, int64_t lastChunk, uint64
 	int64_t targetSize = m_chunkSize;
 	
 	// How long current chunk size would take with the last speed...
-	double msecs = 1000 * targetSize / lastSpeed;
+	double msecs = 1000 * double(targetSize) / lastSpeed;
 	
 	if (msecs < SEGMENT_TIME / 4)
 	{
